requirements.txt
# Core dependencies
selenium==4.16.0
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2

# WebDriver management
webdriver-manager==4.0.1

# Configuration and data handling
python-dotenv==1.0.0
pyyaml==6.0.1
openpyxl==3.1.2
pandas==2.1.4

# Logging and reporting
colorlog==6.8.0
Pillow==10.1.0

# Additional utilities
faker==21.0.0
requests==2.31.0
